requirements.txt
fastapi
aiogram
python-dotenv
uvicorn
loguru